requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.7.2
    # via django
django==4.2.5
    # via -r requirements.in
python-dateutil==2.8.2
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sqlparse==0.4.4
    # via django
typing-extensions==4.7.1
    # via asgiref
